requirements.txt
python-telegram-bot
environs
edge-tts
genanki
openai
requests